+++ python/GafferSceneUI/ShufflePrimitiveVariablesUI.py
@@ -43,13 +43,15 @@ Gaffer.Metadata.registerNode(
 
 	"description",
 	"""
-	ShufflePrimitiveVariables is used to copy or rename arbitrary numbers of primitive variables
-	at the filtered locations. The deleteSource plugs may be used to remove the original source
-	primitive variable(s) after the shuffling has been completed.
+	ShufflePrimitiveVariables is used to copy or rename arbitrary numbers of primitive
+	variables at the filtered locations. The deleteSource plugs may be used to remove
+	the original source primitive variable(s) after the shuffling has been completed.
+	The replaceDestination plugs may be used to specify whether each shuffle should
+	replace already written destination data with the same name.
 
-	An additional context variable `${source}` can be used on the destination plugs to insert
-	the name of each source primitive variable. For example, to append `ref` to all primitive variables
-	set the source to `*` and the destination to `${source}ref`.
+	An additional context variable `${source}` can be used on the destination plugs
+	to insert the name of each source primitive variable. For example, to append `ref`
+	to all primitive variables set the source to `*` and the destination to `${source}ref`.
 	""",
 
 	plugs = {
@@ -59,8 +61,10 @@ Gaffer.Metadata.registerNode(
 			"description",
 			"""
 			The primitive variables to be shuffled - arbitrary numbers of primitive variables
-			may be shuffled via the source/destination plugs. The deleteSource plug may be used
-			to remove the original primitive variable(s).
+			may be shuffled via the source/destination plugs. The deleteSource plug may be
+			used to remove the original primitive variable(s). The replaceDestination plug may
+			be used to specify whether each shuffle should replace already written destination
+			data with the same name.
 			""",
 
 			"divider", True,