+++ src/Gaffer/ShufflePlug.cpp
@@ -46,13 +46,14 @@ using namespace Gaffer;
 
 GAFFER_PLUG_DEFINE_TYPE( ShufflePlug );
 
-ShufflePlug::ShufflePlug( const std::string &source, const std::string &destination, bool deleteSource, bool enabled )
+ShufflePlug::ShufflePlug( const std::string &source, const std::string &destination, bool deleteSource, bool enabled, bool replaceDestination )
 	: ShufflePlug( "shuffle", In, Default | Dynamic )
 {
 	sourcePlug()->setValue( source );
 	destinationPlug()->setValue( destination );
 	deleteSourcePlug()->setValue( deleteSource );
 	enabledPlug()->setValue( enabled );
+	replaceDestinationPlug()->setValue( replaceDestination );
 }
 
 /// Primarily used for serialisation.
@@ -65,6 +66,7 @@ ShufflePlug::ShufflePlug( const std::string &name, Direction direction, unsigned
 	// during `ShufflesPlug::shuffle()`, in order to account for the ${source} variable.
 	addChild( new StringPlug( "destination", direction, "", Plug::Default, IECore::StringAlgo::NoSubstitutions ) );
 	addChild( new BoolPlug( "deleteSource", direction ) );
+	addChild( new BoolPlug( "replaceDestination", direction, true ) );
 }
 
 Gaffer::StringPlug *ShufflePlug::sourcePlug()
@@ -107,6 +109,16 @@ const Gaffer::BoolPlug *ShufflePlug::deleteSourcePlug() const
 	return getChild<BoolPlug>( 3 );
 }
 
+Gaffer::BoolPlug *ShufflePlug::replaceDestinationPlug()
+{
+	return getChild<BoolPlug>( 4 );
+}
+
+const Gaffer::BoolPlug *ShufflePlug::replaceDestinationPlug() const
+{
+	return getChild<BoolPlug>( 4 );
+}
+
 bool ShufflePlug::acceptsChild( const Gaffer::GraphComponent *potentialChild ) const
 {
 	if( !Plug::acceptsChild( potentialChild ) )
@@ -142,6 +154,14 @@ bool ShufflePlug::acceptsChild( const Gaffer::GraphComponent *potentialChild ) c
 		potentialChild->isInstanceOf( BoolPlug::staticTypeId() ) &&
 		potentialChild->getName() == "deleteSource" &&
 		!getChild<Plug>( "deleteSource" )
+	)
+	{
+		return true;
+	}
+	else if(
+		potentialChild->isInstanceOf( BoolPlug::staticTypeId() ) &&
+		potentialChild->getName() == "replaceDestination" &&
+		!getChild<Plug>( "replaceDestination" )
 	)
 	{
 		return true;