+++ Changes.md
@@ -3,6 +3,16 @@
 
 > Note : Python 2 support has been removed. All builds are now using Python 3.
 
+Improvements
+------------
+
+- ShuffleAttributes :
+  - Added `replaceDestination` plugs that may be used to specify whether each shuffle replaces already written destination data with the same name.
+  - Shuffles are now performed in the order they are defined, separate shuffles may write to the same destination.
+- ShufflePrimitiveVariables :
+  - Added `replaceDestination` plugs that may be used to specify whether each shuffle replaces already written destination data with the same name.
+  - Shuffles are now performed in the order they are defined, separate shuffles may write to the same destination.
+
 API
 ---
 
@@ -17,7 +27,7 @@ Breaking Changes
 - Python : Removed support for Python 2.
 - CatalogueUI : Hid OutputIndexColumn from public API.
 
-1.1.x.x ( relative to 1.1.1.0)
+1.1.x.x (relative to 1.1.1.0)
 =======
 
 Improvements